pyproject.toml
[tool.poetry]
name = "spimex-api"
version = "0.1.0"
description = ""
authors = ["John Smith <john@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.1"
ruff = "^0.6.1"
sqlalchemy = "^2.0.32"
pydantic-settings = "^2.4.0"
pytest = "^8.3.2"
pytest-dotenv = "^0.5.2"
python-dotenv = "^1.0.1"
uvicorn = "^0.30.6"
asyncpg = "^0.29.0"
redis = "^5.0.8"
celery = "^5.4.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.8"
pytest-mock = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
